Major rewrite of "postgres_escape" directive.
It is now possible to:
- escape variable in-place,
- escape both anonymous and named regex captures,
- safely re-use same $escaped variable name in different locations,
- concatenate multiple variables/strings and escape them into single variable.

All this comes at the price of not being able to tell the difference between
empty string and non-existing value, which means that starting with this commit
all empty strings will be escaped to NULL value. This behavior can be disabled
by prefixing unescaped value with '=' sign.