Fix compatibility with PostgreSQL 9.x: bytea output.
The issue that hit many (most?) of PostgreSQL adapters is that libpq-8.x
is unable to unescape bytea data from results in text format received
from PostgreSQL-9.x.

But we're always passing bytea data as is to the client (either from
results in binary format when using "postgres_output binary_value"
or escaped from results in text format), so this isn't problem for us.

Patch from Yichun Zhang (agentzh).